Add GraphQL
I had to manually change the schema class and file name to match what Rails expected. My project is named "elasticsearch" but on GitHub, it's "ror-elasticsearch". I intitally changed my module in /config/application.rb from "elasticsearch" to "elasticsearch_" as there was a name collision with the actual Elasticsearch gem files. When I ran the GraphQL installation, this in turned produced a schema class that was named incorrectly as the ending "_" seemed to mess up the naming convention. By renaming the file and schema class, Rails was then able to find and pick up the GraphQL schema properly in GraphiQL.